# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: R&D Intelligence Suite API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: stage
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Project deleted
...
  /formulations:
    get:
      operationId: listFormulations
      tags: [formulations]
      summary: List formulations
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of formulations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Formulation"
    post:
      operationId: createFormulation
      tags: [formulations]
      summary: Create a formulation record
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFormulationRequest"
      responses:
        "201":
          description: Formulation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Formulation"
  /formulations/{id}:
    get:
      operationId: getFormulation
      tags: [formulations]
      summary: Get formulation by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Formulation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Formulation"
    put:
      operationId: updateFormulation
      tags: [formulations]
      summary: Update formulation
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFormulationRequest"
      responses:
        "200":
          description: Updated formulation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Formulation"
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List user notifications
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, manager, scientist, analyst, viewer]
        department:
          type: string
          nullable: true
        avatar:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - role
        - isActive
        - createdAt
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        stage:
          type: string
          enum: [ideation, research, formulation, testing, validation, scale_up, commercialization]
        status:
          type: string
          enum: [active, on_hold, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, critical]
        leadId:
          type: integer
          nullable: true
        startDate:
          type: string
          format: date-time
          nullable: true
        targetDate:
          type: string
          format: date-time
          nullable: true
        successRate:
          type: number
          nullable: true
        revenueImpact:
          type: number
          nullable: true
        productCategory:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        lead:
          $ref: "#/components/schemas/User"
          nullable: true
        taskCount:
          type: integer
        completedTaskCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - stage
        - status
        - priority
        - tags
        - taskCount
        - completedTaskCount
        - createdAt
    CreateProjectRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        stage:
          type: string
          enum: [ideation, research, formulation, testing, validation, scale_up, commercialization]
        status:
          type: string
          enum: [active, on_hold, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, critical]
        leadId:
          type: integer
        startDate:
          type: string
          format: date-time
        targetDate:
          type: string
          format: date-time
        revenueImpact:
          type: number
        productCategory:
          type: string
        tags:
          type: array
          items:
            type: string
      required:
        - name
        - stage
        - status
        - priority
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        stage:
          type: string
          enum: [ideation, research, formulation, testing, validation, scale_up, commercialization]
        status:
          type: string
          enum: [active, on_hold, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, critical]
        leadId:
          type: integer
        startDate:
          type: string
          format: date-time
        targetDate:
          type: string
          format: date-time
        successRate:
          type: number
        revenueImpact:
          type: number
        productCategory:
          type: string
        tags:
          type: array
          items:
            type: string
...
    Formulation:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        version:
          type: string
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/Ingredient"
        sensoryScores:
          $ref: "#/components/schemas/SensoryScores"
          nullable: true
        shelfLifeDays:
          type: integer
          nullable: true
        costPerUnit:
          type: number
          nullable: true
        targetMargin:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, active, approved, rejected]
        createdById:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - name
        - version
        - ingredients
        - status
        - createdAt
    Ingredient:
      type: object
      properties:
        name:
          type: string
        percentage:
          type: number
        costPerKg:
          type: number
          nullable: true
        supplier:
          type: string
          nullable: true
      required:
        - name
        - percentage
    SensoryScores:
      type: object
      properties:
        taste:
          type: number
        texture:
          type: number
        appearance:
          type: number
        aroma:
          type: number
        overall:
          type: number
      required:
        - taste
        - texture
        - appearance
        - aroma
        - overall
    CreateFormulationRequest:
      type: object
      properties:
        projectId:
          type: integer
        name:
          type: string
        version:
          type: string
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/Ingredient"
        sensoryScores:
          $ref: "#/components/schemas/SensoryScores"
        shelfLifeDays:
          type: integer
        costPerUnit:
          type: number
        targetMargin:
          type: number
        notes:
          type: string
        status:
          type: string
          enum: [draft, active, approved, rejected]
      required:
        - projectId
        - name
        - version
        - ingredients
        - status
    UpdateFormulationRequest:
      type: object
      properties:
        name:
          type: string
        version:
          type: string
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/Ingredient"
        sensoryScores:
          $ref: "#/components/schemas/SensoryScores"
        shelfLifeDays:
          type: integer
        costPerUnit:
          type: number
        targetMargin:
          type: number
        notes:
          type: string
        status:
          type: string
          enum: [draft, active, approved, rejected]
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [deadline, update, reminder, mention, system]
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        projectId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - type
        - title
        - message
        - isRead
        - createdAt
...